requirements/common.txt
Flask==1.0.2
Flask-RESTful==0.3.6
Keras==2.2.4
locustio==0.9.0
Pillow==5.3.0
